+++ CHANGELOG.md
@@ -187,6 +187,25 @@ export class IgxCustomFilteringOperand extends IgxFilteringOperand {
   Currently these cover converting submodule imports as well as the deprecation of `igxForRemote` and rename of `igx-tab-bar` to `igx-bottom-nav` from 6.0.0.
 - **Breaking changes**:
     - Removed submodule imports. All imports are now resolved from the top level `igniteui-angular` package. You can use `ng update igniteui-angular` when updating to automatically convert existing submodule imports in the project.
+    - Summary functions for each IgxSummaryOperand class has been made `static`. So now you can use them in the following way:
+    ```typescript
+    import { IgxNumberSummaryOperand, IgxSummaryOperand } from "igniteui-angular";
+    class CustomSummary extends IgxSummaryOperand {
+    constructor() {
+      super();
+    }
+    public operate(data?: any[]) {
+      const result = super.operate(data);
+      result.push({
+        key: "Min",
+        label: "Min",
+        summaryResult: IgxNumberSummaryOperand.min(data)
+      });
+      return result;
+    }
+  }
+    ```
+
 
 ## 6.0.0
 - Theming - You can now use css variables to style the component instances you include in your project.